cvecheck: Add KEV information support
Import Known Exploited Vulnerabilities Catalog[1] data to cve-check result.
Added "KEV" and "KNOWN RANSOMWARE CAMPAIGN USE" keys to support KEV information.
The KEV key takes "Not Found" when a CVE is not in the KEV otherwise
"Found".
The "KNOWN RANSOMWARE CAMPAIGN USE" key takes "Known" when a CVE is in the KEV
otherwise it takes "Unknown". This key is shown when KEV value is "Found".

1: https://www.cisa.gov/known-exploited-vulnerabilities-catalog

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>